+++ core/common/src/main/java/org/openengsb/core/common/workflow/WorkflowService.java
@@ -16,7 +16,10 @@
 
 package org.openengsb.core.common.workflow;
 
+import java.util.Map;
+
 import org.openengsb.core.common.Event;
+
 
 public interface WorkflowService {
     /**
@@ -42,4 +45,17 @@ public interface WorkflowService {
      * @throws WorkflowException when there is a problem while adding the new rule
      */
     void registerFlowTriggerEvent(Event event, String... flowIds) throws WorkflowException;
+
+    /**
+     * Starts a flow with the given id, in the current context's session.
+     * The Objects supplied in the ParameterMap are added to the flow as variables.
+     * 
+     * @return the process' instance ID as returned by drools's KnowledgeSession. It's unique in the scope of the same
+     *         context.
+     * @throws WorkflowException when there is a problem with obtaining the KnowledgeSession or the flow could not be
+     *         started
+     * */
+    long startFlow(String processId, Map<String, Object> parameterMap) throws WorkflowException;
+
 }
+