+++ ui/common/wicket/src/main/java/org/openengsb/ui/common/wicket/model/LocalizableStringModel.java
@@ -1,0 +1,47 @@
+/**
+ * Copyright 2010 OpenEngSB Division, Vienna University of Technology
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.openengsb.ui.common.wicket.model;
+
+import org.apache.wicket.Component;
+import org.apache.wicket.model.LoadableDetachableModel;
+import org.openengsb.core.common.l10n.LocalizableString;
+
+/**
+ * Wicket Model for {@link LocalizableString} objects, used to localize strings coming from other OpenEngSB bundles.
+ *
+ * Example:
+ *
+ * <pre>
+ * add(new Label(&quot;label&quot;, new LocalizableStringModel(this, aLocalizableString)));
+ * </pre>
+ */
+@SuppressWarnings("serial")
+public class LocalizableStringModel extends LoadableDetachableModel<String> {
+
+    private final LocalizableString localizableString;
+    private final Component component;
+
+    public LocalizableStringModel(Component component, LocalizableString localizableString) {
+        this.component = component;
+        this.localizableString = localizableString;
+    }
+
+    @Override
+    protected String load() {
+        return localizableString.getString(component.getSession().getLocale());
+    }
+}