+++ ui/web/src/main/java/org/openengsb/ui/web/LoginPage.java
@@ -28,6 +28,7 @@ import org.openengsb.ui.web.model.User;
 public class LoginPage extends BasePage {
     private User user = new User();
 
+    @SuppressWarnings("serial")
     public LoginPage() {
         Form<User> loginForm = new Form<User>("form") {
             @Override