+++ ui/web/src/main/java/org/openengsb/ui/web/TestClient.java
@@ -20,9 +20,7 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import javax.swing.tree.DefaultMutableTreeNode;
 import javax.swing.tree.DefaultTreeModel;
@@ -54,20 +52,15 @@ import org.apache.wicket.model.LoadableDetachableModel;
 import org.apache.wicket.model.Model;
 import org.apache.wicket.model.PropertyModel;
 import org.apache.wicket.model.ResourceModel;
-import org.apache.wicket.model.StringResourceModel;
 import org.apache.wicket.spring.injection.annot.SpringBean;
 import org.openengsb.core.common.Domain;
 import org.openengsb.core.common.DomainProvider;
 import org.openengsb.core.common.ServiceManager;
-import org.openengsb.core.common.descriptor.AttributeDefinition;
-import org.openengsb.core.common.descriptor.AttributeDefinition.Builder;
 import org.openengsb.core.common.descriptor.ServiceDescriptor;
-import org.openengsb.core.common.l10n.PassThroughStringLocalizer;
 import org.openengsb.core.common.service.DomainService;
-import org.openengsb.ui.web.editor.AttributeEditorUtil;
-import org.openengsb.ui.web.editor.BeanArgumentPanel;
-import org.openengsb.ui.web.editor.fields.AbstractField;
-import org.openengsb.ui.web.model.LocalizableStringModel;
+import org.openengsb.ui.common.wicket.model.LocalizableStringModel;
+import org.openengsb.ui.common.wicket.util.MethodUtil;
+import org.openengsb.ui.web.model.Argument;
 import org.openengsb.ui.web.model.MethodCall;
 import org.openengsb.ui.web.model.MethodId;
 import org.openengsb.ui.web.model.ServiceId;
@@ -102,6 +95,7 @@ public class TestClient extends BasePage {
 
     private ServiceId lastServiceId;
 
+    @SuppressWarnings("serial")
     public TestClient() {
         WebMarkupContainer serviceManagementContainer = new WebMarkupContainer("serviceManagementContainer");
         serviceManagementContainer.setOutputMarkupId(true);
@@ -330,7 +324,7 @@ public class TestClient extends BasePage {
         DefaultMutableTreeNode providerNode =
             new DefaultMutableTreeNode(provider.getName().getString(getSession().getLocale()));
         node.add(providerNode);
-        for (ServiceReference serviceReference : this.services
+        for (ServiceReference serviceReference : services
             .serviceReferencesForDomain(provider.getDomainInterface())) {
             String id = (String) serviceReference.getProperty("id");
             if (id != null) {
@@ -375,28 +369,30 @@ public class TestClient extends BasePage {
     protected void populateArgumentList() {
         argumentList.removeAll();
         Method m = findMethod();
-        List<ArgumentModel> arguments = new ArrayList<ArgumentModel>();
+        List<Argument> arguments = new ArrayList<Argument>();
         call.setArguments(arguments);
         int i = 0;
         for (Class<?> p : m.getParameterTypes()) {
-            ArgumentModel argModel = new ArgumentModel(i + 1, p, null);
+            Argument argModel = new Argument(i + 1, p, null);
             arguments.add(argModel);
-            if (p.isPrimitive() || p.equals(String.class) || p.isEnum()) {
-                Builder builder = AttributeDefinition.builder(new PassThroughStringLocalizer());
-                MethodUtil.addEnumValues(argModel.getType(), builder);
-                builder.id("value").name(
-                    new StringResourceModel("argument", this, new Model<ArgumentModel>(argModel)).getString());
-                AbstractField<?> createEditorField =
-                    AttributeEditorUtil.createEditorField("argument_" + i,
-                        new PropertyModel<String>(argModel, "value"), builder.build());
-                argumentList.add(createEditorField);
-            } else {
-                Map<String, String> beanAttrs = new HashMap<String, String>();
-                argModel.setValue(beanAttrs);
-                argModel.setBean(true);
-                BeanArgumentPanel arg = new BeanArgumentPanel(String.valueOf(i), argModel, beanAttrs);
-                argumentList.add(arg);
-            }
+            MethodArgumentPanel argumentPanel = new MethodArgumentPanel("arg" + i + "panel", argModel);
+            argumentList.add(argumentPanel);
+            // if (p.isPrimitive() || p.equals(String.class) || p.isEnum()) {
+            // Builder builder = AttributeDefinition.builder(new PassThroughStringLocalizer());
+            // MethodUtil.addEnumValues(argModel.getType(), builder);
+            // builder.id("value").name(
+            // new StringResourceModel("argument", this, new Model<Argument>(argModel)).getString());
+            // AbstractField<?> createEditorField =
+            // AttributeEditorUtil.createEditorField("argument_" + i,
+            // new PropertyModel<String>(argModel, "value"), builder.build());
+            // argumentList.add(createEditorField);
+            // } else {
+            // Map<String, String> beanAttrs = new HashMap<String, String>();
+            // argModel.setValue(beanAttrs);
+            // argModel.setBean(true);
+            // BeanEditorPanel arg = new BeanEditorPanel(String.valueOf(i), argModel, beanAttrs);
+            // argumentList.add(arg);
+            // }
             i++;
         }
         call.setArguments(arguments);