+++ CHANGELOG.md
@@ -5,7 +5,10 @@ openengsb-1.1.0.RELEASE
   * [e.g.] Fix problem with setting files
 
 ### New Projects
-  * [e.g.] openengsb-core-taskbox
+  * maven-openengsb-plugin
+  * openengsb-core-taskbox
+  * openengsb-ui-taskbox
+  * openengsb-ui-common-wicket
 
 ### Removed Projects
   * [e.g.] openengsb-domain-scm
@@ -13,12 +16,14 @@ openengsb-1.1.0.RELEASE
 ### Upgraded Projects
  * org.apache.wicket/wicket-*/1.4.13
  * org.neodatis.odb/neodatis-odb/1.9.30.687
+ * junit/junit/4.8.2
+ * org.springframework.security/spring-security-*/3.0.5.RELEASE
 
 ### Remvoed External References
   * [e.g.] javax.ejb/com.springsource.javax.ejb
 
 ### New Features & Changed Behaviour
-  * [e.g.] Added possibility to delete workfows
+  * Added tooling plugin to execute license-check plugin and assemble scripts via mvn.
 
 ### Depricated or Removed Features
   * [e.g.] org.openengsb.domain.scm.doSomething() is removed